Move get_next_strobe() to unused.cpp
